fastapi/requirements.txt
fastapi==0.109.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
typer[all]==0.9.0
uvicorn==0.26.0
pymysql==1.1.0
SQLAlchemy==2.0.25
alembic==1.13.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0